Prepare for field based datamanagers
