+++ src/senaite/core/datamanagers/content/configure.zcml
@@ -1,0 +1,15 @@
+<configure xmlns="http://namespaces.zope.org/zope">
+
+  <!-- Data Manager for Analysis Services -->
+  <adapter factory=".analysisservice.AnalysisServiceDataManager" />
+
+  <!-- Data Manager for Analyses -->
+  <adapter factory=".sample.SampleDataManager" />
+
+  <!-- Data Manager for Analyses -->
+  <adapter factory=".analysis.RoutineAnalysisDataManager" />
+
+  <!-- Data Manager for Reference Analyses -->
+  <adapter factory=".analysis.ReferenceAnalysisDataManager" />
+
+</configure>
